+++ web/apps/photos/src/components/Collections/CollectionShare/publicShare/EnablePublicShareOptions.tsx
@@ -1,6 +1,8 @@
-import { MenuItemDivider } from "@/new/common/components/menu/MenuItemDivider";
-import { MenuItemGroup } from "@/new/common/components/menu/MenuItemGroup";
-import { MenuSectionTitle } from "@/new/common/components/menu/MenuSectionTitle";
+import {
+    MenuItemDivider,
+    MenuItemGroup,
+    MenuSectionTitle,
+} from "@/new/common/components/Menu";
 import { EnteMenuItem } from "@ente/shared/components/Menu/EnteMenuItem";
 import DownloadSharp from "@mui/icons-material/DownloadSharp";
 import LinkIcon from "@mui/icons-material/Link";
