+++ web/apps/photos/src/components/PhotoViewer/ImageEditorOverlay/CropMenu.tsx
@@ -1,5 +1,4 @@
-import { MenuItemGroup } from "@/new/common/components/menu/MenuItemGroup";
-import { MenuSectionTitle } from "@/new/common/components/menu/MenuSectionTitle";
+import { MenuItemGroup, MenuSectionTitle } from "@/new/common/components/Menu";
 import { EnteMenuItem } from "@ente/shared/components/Menu/EnteMenuItem";
 import CropIcon from "@mui/icons-material/Crop";
 import { t } from "i18next";
